backend/requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
email-validator==1.1.3
python-dotenv==0.19.0
SQLAlchemy==1.4.23
Werkzeug==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Werkzeug==2.0.1
click==8.0.1
SPARQLWrapper==2.0.0
rdflib==6.3.2
flask-cors==3.0.10
